Have `wait_for_download_completion` initialize the path in the same way as the Realm constructor
`RealmClass<T>::wait_for_download_completion` was initializing the path
before calling `SyncClass<T>::populate_sync_config`, which resulted in
any custom path it specified being overwritten. It was also treating an
unspecified path as indicating the default Realm path should be used,
when synced Realms use a path derived from the user and URL.

This now results in `Realm.openAsync` raising the
`IncompatibleSyncedRealmError` exception at the correct time since
the correct path is now used when the Realm is initially opened. To
accommodate this, `openAsync` is updated to catch exceptions and forward
them to its callback.