Don't close the Realm until the JavaScript download completion callback has been run.
The C++ download completion callback runs on a background thread, and
bounces over to the main thread to invoke the JavaScript callback. By
closing the Realm on the background thread we would often close it
before the JavaScript callback had a chance to open it, negating our
attempts to keep the Realm open long enough for the JS callback to reuse
its sync session.