Check for REALM_DISABLE_ANALYTICS earlier in submit-analytics.js
Node's `require()` is synchronous and somewhat slow, so importing the modules
for submit-analytics.js ends up being about half of the runtime of
`require('realm')`. With this change, when analytics is disabled, nearly all of
the time spent on `require('realm')` (~100ms on my machine) is spent in
node-pre-gyp-related functionality for loading the native module.