Expose disable_sync_to_disk() via an env variable
This makes core not actually flush data to disk when durability is not needed,
such as when running tests. Doesn't help much for realm-js since the bulk of
the time in them is spent waiting for things to time out, but it cuts the run
time of the ROS tests in half.