Update Readme.md to include debugging disclaimer
Some users on #491 have now reported investing hours of dev time to integrate Realm before having to remove it altogether due to the debugger being too slow for their team or project to use. We need to provide an honest disclaimer to developers until this issue is fixed as Realm is breaking a core, built-in piece of React Native functionality.