Improve handling of parallel CI jobs.
1. Sometimes, it is necessary to run multiple parallel CI jobs on the same
physical machine w/o any virtualization or containerization (a good
example of that is building an iOS app). To work properly in that
environment, realm should download a temp file to it's own unique temp
folder to avoid clashes (that are actually fail with `symlink` syscall
failing with `EEXISTS`).

2. To avoid an ever-growing temp folder, a new environment variable was
introduced, that you can set to be inside a job's folder. That way, CI
will be responsible for cleaning up after itself and keeping only the
necessary amount of files.

Theoretically, all of that can be solved by overriding `TMP` env
variable but in practice in many-many environments it breaks the other
part of toolchain (even Xcode builds themselves sometimes).