Check the correct name when adding permissions object schemas to the schema
The `name` property of the schema object is the name of the class (i.e. `Role`
for the role object) rather than the schema name (i.e. `__Role`). This happened
to work if only the predefined schema objects were used and no user-defined
classes had the non-underscore-prefixed names, but broke otherwise.