Change how the Electron test process is torn down
There were some race conditions where we would incorrectly think that the test
process had exited unexpectedly due to that it could complete successfully
while we were in the middle of processing logs and never see the last log
lines. In addition, when exiting immediately after logging some log messages
seem to be lost, so wait for the parent process to tell the child to exit.