Fix a race when tearing down out-of-process notifications
Any change objects passed to an out-of-process notification worker while the
worker was shutting down did not release the native ChangeObject, resulting in
it persisting until the JS object was garbage collected. This probably did not
matter for any real users, but it resulted in the tests occasionally crashing
due to a Realm still being open after it should have been closed.