Make AppClass::transport_generator initialized at const-init time (#3665)
This fixes the "bad_function_call" issue where it was invoked while null
(probably actually uninitialized). I'm not sure why it wasn't
initialized yet, since the compiler should have ensured that it was.
There is likely a compiler bug there but I'll need to dig in more. To
work around that, I made it a simple function pointer, which lambdas can
implicitly convert to, and all of that should be possible at const-init
time (basically, it should be burned into the binary image with nothing
happening at runtime). Unfortunately due to a different compiler bug,
(https://developercommunity.visualstudio.com/t/const-init-of-function-pointers-from-lambdas/1383098),
this requires putting a `+` in front of the lambda to explicitly decay
it to a function pointer, rather than relying on the implicit conversion.