Fix Pipelines
There seems to be a problem with the android-emulator-runner action.  Using the workaround posted in this issue:
https://github.com/ReactiveCircus/android-emulator-runner/issues/168

The linting pipeline is running ubuntu-latest, which does not have cocoa pods.  One of the lerna dependencies (realm-react-native-tests) was invoking `pod install`.  Changed this to use the npm package `pod-install` which will not crash on non-mac systems.