send App, User, and Client context headers (realm-web)
Extra HTTP headers will be sent if Realm.App or Realm.User objects
have a headers property or a dynamic getter function defined
on them, or if headers are given as an option to Realm.User.mongoClient.
Priority cascades from App to User to mongoClient (later replaces
earlier in this list).