Create github workflow to build and publish jsdocs (#3762)
The workflow is triggered for PRs that change documentation files as well as for
any release tags (vX.Y.Z).
Additionally for tags the documentation is published to an S3 bucket under
a direcotry with the version as its name.

For uploading to work, three github secrets need to be configured: DOCS_S3_ACCESS_KEY,
DOCS_S3_SECRET_KEY and DOCS_S3_BUCKET_NAME