Remove Examples Folder (#4183)
The contents of this folder are two old unmatinained projects that have probably outlived their purpose.  If it is desired they stick around they should be updated and moved to either their own repo or into packages.  I think the `examples` folder will only confuse users into believing there is an example here to be used.