Split JSDoc workflow and fix uploading (#4277)
The current workflow uploads the docs to the wrong location for archival, and does not update the copy of the docs at /latest. This change splits the workflow in two - one which builds the JSDocs for every pull request but does not publish them, and one which only runs on vX.Y.Z tag builds, and uploads the docs both to /latest and for archival.