Explicitly flush the React Native UI task queue whenever we call from C++ to JS
React Native has its own JS "microtask" queue, implemented in https://github.com/facebook/react-native/blob/main/Libraries/Core/Timers/JSTimers.js. Any asynchronous work, e.g. `setTimeout` or `setImmediate`, is added to this queue, and the queue is flushed whenever a message is sent from React Native's native core to JS: https://github.com/facebook/react-native/blob/main/Libraries/BatchedBridge/MessageQueue.js#L108.

However, our native to JS messages are not being passed via this abstraction - instead, we hook directly into the JS engine. This means that React Native is not aware that when Realm has done some async work, we might need to update the UI (and therefore flush the task queue), and this can result in Realm-related UI updates not showing until some action is taken which causes React Native to send a message from the core to JS (e.g. touching the screen), which flushes this task queue, resolving pending promises and updating the UI.

Two solutions were considered for whenever Realm has performed some work which might require a UI update:
1. Make Realm JS "play by the rules" that React Native expects native code to use, and e.g. send a dummy message via React Native whenever we have done work.
2. Expose the internal `callReactNativeMicrotasks` function from React Native in a way that we can directly call that from C++ whenever we have done work.

We decided to go with option 2, as it is simpler to implement and we felt we can implement it in a reasonably clean way using the Realm constructor. This does mean we are depending on React Native internal implementation details, but as we already are potentially quite tightly coupled to RN versions, we didn't feel this was too much of an issue.