React Integration (#4197)
* Add launch command for jest

* Documentation of Realm React
Add tsdoc notation to the public facing interfaces to realm/react.
Also add comments to any possibly unclear blocks of code

* Rearrange tsdocs for hooks and provider
If one ctrl-clicks on the hooks or the provider in vscode, it navigates
to the interface of createRealmContext.  The comments have been moved here
so the documentation can be found faster.

* Add render tests to prove modifications ensure no unnecessary re-renders.
* Collection Cache with Strong References (#4209)
* Replace extra data re-render mechanism by just returning a new empty proxy

* Support sorted and filter
Deletions will cause the cache to be cleared.  We have to live this this
until the listener callbacks provide the ability to retrieve the objectId
of the deleted object.

* Refactor Test Suite
Make use of description.each to test different types of collections.
This includes sorted and filtered to show failing tests that must be
fixed in the implementation of `cachedCollection`.

* Add ability to debug the current open js file using node.
* Support List property in useObject (#4274)

* Phase 2: Add cache support for Realm.List in Realm.Object
Introduces the cachedObject which adds a proxy around a Realm.Object
to handle re-rendering on changes and to cache references to all
Realm.Objects within a Realm.List.

* Fix types for Object Listener
When a class extends Realm.Object, adding a listener to an instance of the
object would not correctly type the object returned in the callback.
With this change, the callback returns an object of the same type as the
instance it was called from.

* Refactor useObject render tests into a single test suite
* Rename example to testApp and remove old code.
* Add tests for lists within a collection