Fix intermittent failures in flexible sync test runs by always awaiting synchronisation.
See comment for further details - it wasn't possible to find an alternate resolution after a few days debugging, and the issue does not appear to affect non-test code.

The original problem could be reproduced by running two tests in a loop (i.e. create a for loop around the describe blocks), one of which updates the subs without waiting for synchronisation, and the other of which does wait, for example "if a subscription with the same query is added, the old reference can be removed" and "removes all subscriptions and returns the number of subscriptions removed".