Realm js release workflows (#4588)
* release workflows

* prepare templates release workflow

* prepare-release: take 'tag' as a manual workflow input

Pass it as npm publish --tag
Mark the release as a prerelease if tag is not 'latest'

* prepare-release workflow: Add PAT token to the Create PR step

Supplying a PAT token instead of the default GitHub token means that
GitHub will run the PR workflows on the newly created PR.

By default it doesn't do that, to avoid creating endless loops of workflows.

* publish-release workflow: Continue if release-to-slack fails

* Get the pkg version without quotes around it

npm pkg get version
produces "1.19.5" instead of just 1.19.5, causing problems later.

* prepare-release workflow: Don't mark the new release PR as draft

* Bugfix: prepare-release workflow: update version number in package.json

The regex for the filename was wrong.

Replaced the action with jq, which is more accurate in replacing only
the top level version: key.

* publish-release workflow: Bugfix in Upload Github Release artifacts

Install npm prebuild before trying to use it.

* Fixes to 'Create vNext PR'

* release workflow: Move all build products into place for uploading.

* Remove temporary publish job from PR workflow, moved to release workflow

* Don't put release artifacts on the github release page

They are published in the npm release.

* Add workflow to publish templates

Co-authored-by: Richard Braakman <rebraakman@gmail.com>