Optimize Realm.Object accessor creation
This contains a handful of changes primarily focused on optimizing the
creation of instances derived from Realm.Object. Note that this is the binding
of a JS object to an existing DB object, not just the DB insert path. There
are also a handful of correctness changes in related areas in order to avoid
the potential for conflicts with users' field names.

This improves our link traversal integration test ("Car") by 3.5x over status
quo and makes us now 2.5x faster than v11. This test has some noise, but on my
machine I was seeing around {before: 200k ops/s, after: 700k, v11: 350k}.

Changes:
* The `realm` instance member is now a unique REALM symbol field bound into
  the synthetic prototype[1].
* `keys()` method changes:
  * Internal consumers now use symbol properties `this[KEY_ARRAY]` or
    `this[KEY_SET]` depending on whether they need O(1) membership checking.
  * Both new properties live on the synthetic prototype.
  * `keys()` is now a normal method on Realm.Object rather than being on the
    synthetic constructor.
  * `keys()` now returns a copy of the array to prevent caller mutation of the
    "true" list of keys.
* We no longer use `Object.createProperty()` to mark our internal properties
  as not writable or enumerable. It is a very slow method at least on V8.
  Since we are already using a Proxy, we just modify the result returned by
  `getOwnPropertyDescriptor()` from the proxy handler instead.
* We now lazily create the `ObjectListeners` for each object when the first
  listener is added. This avoids a rather expensive operation for objects that
  never have any listeners registered.
* Use `declare` for internal properties we synthetically inject. This tells
  tsc not to generate a class field declaration when using a new enough target
  and has weird interactions with some of the techniques we are using.

[1] This is for the synthetic constructor which we create in ClassMap to be
the most-derived class for all instances of Realm.Object that we create. This
is also where the get/set properties for user fields live. We create a unique
synthetic constructor for each combination of user type + `Realm` instance.