OrderedCollection no longer requires a Results
Previously it always held a `Results` object, even for the `List` and `Set`
subclasses. That was expensive to construct, and possibly changed the runtime
complexity. Now it will directly work with whatever native object it's
subclass represents. This required a small change to the core spec file to
give Results and Collections a compatible API.