[useObject] Update primaryKeyRef on change (#5620)
Within the `useObject` hook, the internal ref `primaryKeyRef` was never updated, causing successive calls to `arePrimaryKeysIdentical` to fail.
Issue 1:
If the hook is called with a primary key `a`, then again with `b`, and then again with `a`, it will return the object for the primary key `b`.
Issue 2:
After a primary key change, `arePrimaryKeysIdentical` will always return false, potentially causing the cached object to be recreated on each render.